[fixed] City radius
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk@2545 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ MMOCoreORB/src/server/zone/managers/city/CityManagerImplementation.cpp
@@ -16,7 +16,7 @@
 #include "server/zone/objects/region/Region.h"
 
 Vector<uint8> CityManagerImplementation::citizensPerRank;
-Vector<uint8> CityManagerImplementation::radiusPerRank;
+Vector<uint16> CityManagerImplementation::radiusPerRank;
 int CityManagerImplementation::cityUpdateInterval = 0;
 int CityManagerImplementation::newCityGracePeriod = 0;
